requirements.txt
asgiref==3.2.10
Django==3.1
django-appconf==1.0.4
django-compressor==2.4
django-crispy-forms==1.9.0
django-markdownx==3.0.1
django-redis==4.12.1
django-sass-processor==0.8.2
django-select2==7.6.2
libsass==0.20.1
Markdown==3.3.3
Pillow==8.0.1
pytz==2020.4
rcssmin==1.0.6
redis==3.5.3
rjsmin==1.1.0
six==1.15.0
sqlparse==0.4.1
